refactor: estrai domain/(layer,path_resolver) da cli/
Sposta layer_ops.py e path_resolver (da _helpers) in toolkit.domain/.
cli/layer_ops.py diventa re-export puro.
MCP e test aggiornati a importare da domain/.
Azzera dipendenza MCP→cli/ per questo modulo.

10 file, 0 logica modificata, solo import path.
